Add response interceptor to refresh tokens on 401
This covers cases where an access-token might be invalidated before it
expires

Renames existing interceptor to note it's a request one, as well as the
file to a generic 'interceptors' based one.
