+++ src/lib/utils/api/index.ts
@@ -1,2 +1,2 @@
 export * from './callSnykApi';
-export * from './refreshTokenInterceptor';
+export * from './interceptors';
